server/requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==41.0.4
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
Pillow==10.0.1
pycparser==2.21
PyJWT==2.7.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
